+++ packages/react/src/components/combobox/PlCombobox.tsx
@@ -653,7 +653,12 @@ export function PlCombobox<Multiple extends boolean | undefined = false>({
 
       {description ? (
         <Field.Description
-          className={cx(metaTextClasses[size], 'text-(--plass-muted-fg)', classNames?.description)}
+          className={cx(
+            'm-0',
+            metaTextClasses[size],
+            'text-(--plass-muted-fg)',
+            classNames?.description
+          )}
         >
           {description}
         </Field.Description>