+++ packages/react/src/components/number-field/PlNumberField.tsx
@@ -411,7 +411,12 @@ export function PlNumberField({
 
       {description ? (
         <Field.Description
-          className={cx(metaTextClasses[size], 'text-(--plass-muted-fg)', classNames?.description)}
+          className={cx(
+            'm-0',
+            metaTextClasses[size],
+            'text-(--plass-muted-fg)',
+            classNames?.description
+          )}
         >
           {description}
         </Field.Description>