+++ packages/react/src/components/text-field/PlTextField.tsx
@@ -327,7 +327,12 @@ export const PlTextField = /* @__PURE__ */ React.forwardRef<
 
       {description ? (
         <Field.Description
-          className={cx(metaTextClasses[size], 'text-(--plass-muted-fg)', classNames?.description)}
+          className={cx(
+            'm-0',
+            metaTextClasses[size],
+            'text-(--plass-muted-fg)',
+            classNames?.description
+          )}
         >
           {description}
         </Field.Description>